python-approach/requirements.txt
azure-core==1.9.0
azure-functions==1.5.0
azure-storage-blob==12.6.0
easyocr==1.4.1
pdf2image==1.16.0
spacy
pytesseract
pandas
concise-concepts